requirements-dev.txt
asgiref==3.6.0
dj-database-url==1.3.0
dj-static==0.0.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.4.3
static3==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
